Api error handling improvements

+++ Assets/ElrondUnityTools/Scripts/ModifiedScripts/TranscationData.cs
@@ -22,8 +22,8 @@ namespace ElrondUnityTools
         public string sender;
         public string receiver;
         public string value;
-        public int gasPrice;
-        public int gasLimit;
+        public long gasPrice;
+        public long gasLimit;
         public string data;
         public string chainId;
         public int version;
@@ -35,8 +35,8 @@ namespace ElrondUnityTools
             sender = tx.from;
             receiver = tx.to;
             value = tx.amount;
-            gasPrice = int.Parse(tx.gasPrice);
-            gasLimit = int.Parse(tx.gasLimit);
+            gasPrice = long.Parse(tx.gasPrice);
+            gasLimit = long.Parse(tx.gasLimit);
             data = Convert.ToBase64String(Encoding.UTF8.GetBytes(tx.data));
             chainId = tx.chainId;
             version = tx.version;
